apple_bot/requirements.txt
aiogram==3.24.0
python-dotenv==1.2.1
SQLAlchemy==2.0.46
aiosqlite==0.22.1